Add PolicyRules file
The PolicyRules file is for the Policy Analyzer from Microsoft. With this tool it is possible to read out and compare local registry and local policy to a defined baseline. This PolicyRule file contains all the rules that are needed to check the Group Policy and Registry settings that are defined in the Windows 10 Hardening checklist.