Update translation for auditpol policies
I tried to start an audit on my computer, after an hailmary application, to see my Hardening score, and I figured out that some policies that I changed did not match the recommended value. 
When the method retrieves the information, it retrieves it in the system language, so, if the csv file isn’t in the same language, the values don’t match.